+++ semana12/labeddit/src/globalstate/GlobalState.js
@@ -6,6 +6,7 @@ import { BASE_URL,headers } from '../constants/url'
 
 const GlobalState = (props) => {
 
+const [comments,setComments] = useState([{}])
 const [voted,setVoted] = useState(false)
 const [posts,setPosts] = useState([{}])
 const [idPost,setIdPost] = useState([{}])
@@ -20,6 +21,20 @@ const getPost = () => {
     .catch((err) => {
         alert(err.message)
     })
+}
+
+const getComments = () => {
+    if(idPost.length > 1){
+        axios
+        .get(`${BASE_URL}/posts/${idPost}/comments`,headers)
+        .then((response) => {
+            setComments(response.data)
+            console.log(response.data)
+        })
+        .catch((err) => {
+            alert(err.message)
+        })
+    }
 }
 
 
@@ -43,7 +58,7 @@ useEffect(() => {
 
 
 return (
-    <GlobalStateContext.Provider value = {{posts,idPost,setIdPost,DeleteVote,setVoted,getPost}}>
+    <GlobalStateContext.Provider value = {{posts,idPost,comments,setIdPost,DeleteVote,setVoted,getPost,getComments}}>
         {props.children}
     </GlobalStateContext.Provider>
 )
